+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/DocumentMediaGroupContent.kt
@@ -1,8 +1,12 @@
 package dev.inmo.tgbotapi.types.message.content.media
 
 import dev.inmo.tgbotapi.types.InputMedia.*
+import dev.inmo.tgbotapi.types.files.DocumentFile
+import dev.inmo.tgbotapi.types.files.abstracts.TelegramMediaFile
 import dev.inmo.tgbotapi.types.message.content.media.MediaGroupContent
 
 interface DocumentMediaGroupContent : MediaGroupContent {
+    override val media: DocumentFile
+
     override fun toMediaGroupMemberInputMedia(): DocumentMediaGroupMemberInputMedia
 }